backend/requirements.txt

fastapi
uvicorn
python-dotenv
openpyxl
pydantic
strands-agents
strands-agents-bedrock
